+++ backend/app/Models/User.php
@@ -23,6 +23,7 @@ class User extends Authenticatable implements FilamentUser
         'password',
         'is_superadmin',
         'product_tour_completed_at',
+        'checklist_dismissed_at',
     ];
 
     protected $hidden = [
@@ -37,6 +38,7 @@ class User extends Authenticatable implements FilamentUser
             'password' => 'hashed',
             'is_superadmin' => 'boolean',
             'product_tour_completed_at' => 'datetime',
+            'checklist_dismissed_at' => 'datetime',
         ];
     }